Add files via upload
Each of the teams had one technology student and several culinary students. The task of the technology students was to listen and document requirements brainstorming (but without proposing any solutions or ideas of their own, which can be difficult for new engineers). This directory contains all of the fruits of their labor!

The name of the team comes from the color of the shirt being worn by the Technology student gathering the requirements